fix: fix the telemetry package write errors

+++ src/PortingAssistantExtensionTelemetry/Model/SolutionMetrics.cs
@@ -7,6 +7,6 @@ namespace PortingAssistantExtensionTelemetry.Model
     public class SolutionMetrics : MetricsBase
     {
         public string SolutionPath { get; set; }
-        public int AnalysisTime { get; set; }
+        public double AnalysisTime { get; set; }
     }
 }
